+++ EPPlusTest/FormulaParsing/Excel/Functions/Math/RoundDownTests.cs
@@ -24,7 +24,6 @@
 *
 * For code change notes, see the source control history.
 *******************************************************************************/
-using System;
 using EPPlusTest.FormulaParsing.TestHelpers;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OfficeOpenXml.FormulaParsing.Excel.Functions.Math;
@@ -284,6 +283,51 @@ namespace EPPlusTest.FormulaParsing.Excel.Functions.Math
 			Assert.AreEqual(12.3d, booleanTrue.Result);
 			Assert.AreEqual(12d, booleanFalse.Result);
 		}
-		#endregion	
+
+		[TestMethod]
+		public void RounddownShouldReturnCorrectResultWithPositiveNumber()
+		{
+			var func = new Rounddown();
+			var args = FunctionsHelper.CreateArgs(9.999, 2);
+			var result = func.Execute(args, this.ParsingContext);
+			Assert.AreEqual(9.99, result.Result);
+		}
+
+		[TestMethod]
+		public void RounddownShouldHandleNegativeNumber()
+		{
+			var func = new Rounddown();
+			var args = FunctionsHelper.CreateArgs(-9.999, 2);
+			var result = func.Execute(args, this.ParsingContext);
+			Assert.AreEqual(-9.99, result.Result);
+		}
+
+		[TestMethod]
+		public void RounddownShouldHandleNegativeNumDigits()
+		{
+			var func = new Rounddown();
+			var args = FunctionsHelper.CreateArgs(999.999, -2);
+			var result = func.Execute(args, this.ParsingContext);
+			Assert.AreEqual(900d, result.Result);
+		}
+
+		[TestMethod]
+		public void RounddownShouldReturn0IfNegativeNumDigitsIsTooLarge()
+		{
+			var func = new Rounddown();
+			var args = FunctionsHelper.CreateArgs(999.999, -4);
+			var result = func.Execute(args, this.ParsingContext);
+			Assert.AreEqual(0d, result.Result);
+		}
+
+		[TestMethod]
+		public void RounddownShouldHandleZeroNumDigits()
+		{
+			var func = new Rounddown();
+			var args = FunctionsHelper.CreateArgs(999.999, 0);
+			var result = func.Execute(args, this.ParsingContext);
+			Assert.AreEqual(999d, result.Result);
+		}
+		#endregion
 	}
 }
